requirements.txt
pandas
torch
transformers
datasets
numpy
scikit-learn
sentencepiece
tqdm
flask
pytest
matplotlib
seaborn
PyYAML